Add correct log file

+++ src/Daemon/Program.cs
@@ -1,4 +1,6 @@
-﻿using System.IO;
+﻿using System;
+using System.IO;
+using Common;
 using Microsoft.Extensions.Logging;
 
 namespace Daemon;
@@ -7,16 +9,23 @@ class Program
 {
     static ILoggerFactory CreateLoggerFactory()
     {
-        // Define the path to the text file
-        // TODO: switch to correct file
-        var logFilePath = "gcron.log";
 
         // Create a StreamWriter to write logs to a text file
-        var logFileWriter = new StreamWriter(logFilePath, append: true);
         return LoggerFactory.Create(builder =>
         {
-            // Add a custom log provider to write logs to text files
-            builder.AddProvider(new FileLoggerProvider(logFileWriter));
+            // Define the path to the text file
+            var logFilePath = Settings.LogFileLocation;
+
+            if (File.Exists(logFilePath))
+            {
+                var logFileWriter = new StreamWriter(logFilePath, append: true);
+                // Add a custom log provider to write logs to text files
+                builder.AddProvider(new FileLoggerProvider(logFileWriter));
+            }
+            else
+            {
+                Console.Error.WriteLine($"Log file ({logFilePath}) doesn't exists (or insufficient privileges), no logs will be generated");
+            }
         });
     }
